Refactor root API to health for improved clarity
- Rename any_auth/api/root.py to any_auth/api/health.py to better reflect
  its purpose as a health check endpoint
- Update imports and router registration in build_app.py to use
  health_router instead of root_router

This change improves code clarity and maintainability by making the
endpoint's intent explicit.
